[2/2] Rename controller class
Rename module
